requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dataclasses==0.6
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
idna-ssl==1.1.0
importlib-metadata==4.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
pyee==7.0.4
slackclient==2.9.3
slackeventsapi==2.2.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.4.1
